Purge TechnicianOrderRepository
