+++ Core/Repositories/IOrderRepository.cs
@@ -1,10 +1,12 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
 using Warehouse.Core.Models;
 
 namespace Warehouse.Core.Repositories
 {
     public interface IOrderRepository
     {
+        Task<IEnumerable<Order>> GetOrders(int technicianId);
         Task<Order> GetOrder(int id);
     }
 }
