+++ Persistence/Repositories/OrderRepository.cs
@@ -1,4 +1,6 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Warehouse.Core.Models;
 using Warehouse.Core.Repositories;
@@ -12,6 +14,16 @@ namespace Warehouse.Persistence.Repositories
         public OrderRepository(ApplicationDbContext context)
         {
             _context = context;
+        }
+
+        public async Task<IEnumerable<Order>> GetOrders(int technicianId)
+        {
+            return await _context.Orders
+                .Include(order => order.OrderDetails)
+                .ThenInclude(orderDetail => orderDetail.Product)
+                .ThenInclude(product => product.Merchant)
+                .Where(order => order.TechnicianId == technicianId)
+                .ToListAsync();
         }
 
         public async Task<Order> GetOrder(int id)
